backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9

pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0

prophet==1.1.5
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

celery[redis]==5.3.6
redis==5.0.4

python-dotenv==1.0.1
httpx==0.27.0
anthropic>=0.40.0
slowapi==0.1.9

# Reports
reportlab==4.2.0
openpyxl==3.1.2

# PDF text extraction
pdfplumber==0.11.4

# Gmail OAuth
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0

# LangGraph + RAG
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
